Minor tweaks to avg_word_length. Hopefully we are making this a little more varied over time.
New TODO:
We have "name" morphemes and stuff, use those as well in the generation of nouns and objects.
While this entire project is just to generate a basically meaningless language, the language itself should at least try to be similar in nature to real languages.
So it needs some names and places to reference.....!!